add login request at like btn

+++ src/components/common/VideoInf.vue
@@ -22,6 +22,7 @@
 <script>
 // import lodash from "lodash";
 import { mapState } from "vuex";
+import router from "@/router";
 
 export default {
   name: "VideoInf",
@@ -57,30 +58,34 @@ export default {
   methods: {
     // 찜 버튼 누르기
     like(id) {
-      this.$store.dispatch("getLiked").then(() => {
-        let flag = true;
-        for (let video of this.liked) {
-          if (video.id === id) {
-            flag = false;
-            break;
-          }
-        }
-        if (flag) {
-          this.$store.dispatch("createLiked", { id: id });
-          alert("찜 리스트에 추가되었습니다.");
-        } else {
-          this.$store.dispatch("deleteLiked", { id: id });
-          let idx = 0;
-          for (let photo of this.shownPhotos) {
-            if (photo.id === id) {
-              this.shownPhotos.splice(idx, 1);
+      if (localStorage.getItem("access-token") === null) {
+        alert("로그인이 필요한 서비스입니다.");
+        router.push({ name: "userlogin" });
+      } else
+        this.$store.dispatch("getLiked").then(() => {
+          let flag = true;
+          for (let video of this.liked) {
+            if (video.id === id) {
+              flag = false;
               break;
             }
-            idx++;
           }
-          alert("찜 리스트에서 삭제되었습니다.");
-        }
-      });
+          if (flag) {
+            this.$store.dispatch("createLiked", { id: id });
+            alert("찜 리스트에 추가되었습니다.");
+          } else {
+            this.$store.dispatch("deleteLiked", { id: id });
+            let idx = 0;
+            for (let photo of this.shownPhotos) {
+              if (photo.id === id) {
+                this.shownPhotos.splice(idx, 1);
+                break;
+              }
+              idx++;
+            }
+            alert("찜 리스트에서 삭제되었습니다.");
+          }
+        });
     },
     // 사진 가져와서 출력
     getPhotos: function () {
@@ -214,11 +219,13 @@ export default {
 }
 
 .card img {
+  border-radius: 7px;
   width: 100%;
   transition: 0.3s;
 }
 
 .card:hover img {
+  border-radius: 12px;
   width: 100%;
   filter: brightness(35%);
   transition: 0.3s;
